Removes TestLifecycleAwareListener and call TestLifecycleAware methods in per test listener
