+++ modules/kotest/src/main/kotlin/org/testcontainers/containers/StartablePerTestListener.kt
@@ -6,6 +6,9 @@ import io.kotest.core.test.TestResult
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import org.testcontainers.lifecycle.Startable
+import org.testcontainers.lifecycle.TestDescription
+import org.testcontainers.lifecycle.TestLifecycleAware
+import java.util.*
 
 
 /**
@@ -22,13 +25,38 @@ internal class StartablePerTestListener(private val startable: Startable) : Test
 
     override suspend fun beforeTest(testCase: TestCase) {
         withContext(Dispatchers.IO) {
+            runBeforeTestForTestLifecycleAware(testCase)
             startable.start()
         }
     }
 
     override suspend fun afterTest(testCase: TestCase, result: TestResult) {
         withContext(Dispatchers.IO) {
+            runAfterTestForTestLifecycleAware(testCase, result)
             startable.stop()
         }
     }
+
+    private fun runBeforeTestForTestLifecycleAware(testCase: TestCase) {
+        startable.toTestLifecycleAware()?.beforeTest(testCase.toTestDescription())
+    }
+
+    private fun runAfterTestForTestLifecycleAware(testCase: TestCase, result: TestResult) {
+        startable.toTestLifecycleAware()?.afterTest(testCase.toTestDescription(), Optional.ofNullable(result.error))
+    }
+
+    private fun Startable.toTestLifecycleAware() = this as? TestLifecycleAware
 }
+
+
+private fun TestCase.toTestDescription() = object : TestDescription {
+
+    override fun getFilesystemFriendlyName(): String {
+        return name.replace(" ", "-")
+    }
+
+    override fun getTestId(): String {
+        return description.id()
+    }
+}
+
