+++ modules/kotest/src/test/kotlin/org/testcontainers/containers/TestStartable.kt
@@ -2,7 +2,7 @@ package org.testcontainers.containers
 
 import org.testcontainers.lifecycle.Startable
 
-internal class TestStartable : Startable {
+internal open class TestStartable : Startable {
     var startCount = 0
     override fun start() {
         startCount++
